third_party/WebKit/LayoutTests/fast/frames/sandboxed-iframe-domain-expected.txt

CONSOLE ERROR: line 1: Uncaught SecurityError: Failed to set the 'domain' property on 'Document': assignment is forbidden for sandboxed iframes.
